+++ lib/src/tooltip/tooltip_wrapper.dart
@@ -66,6 +66,7 @@ class ToolTipWrapper extends StatefulWidget {
     required this.tooltipPadding,
     required this.toolTipSlideEndDistance,
     required this.targetTooltipGap,
+    this.semanticEnable = false,
     this.scaleAnimationAlignment,
     this.tooltipPosition,
     this.titlePadding,
@@ -108,6 +109,7 @@ class ToolTipWrapper extends StatefulWidget {
   final EdgeInsets targetPadding;
   final ShowcaseController showcaseController;
   final double targetTooltipGap;
+  final bool semanticEnable;
 
   @override
   State<ToolTipWrapper> createState() => _ToolTipWrapperState();
@@ -179,7 +181,9 @@ class _ToolTipWrapperState extends State<ToolTipWrapper>
                 : SystemMouseCursors.click,
             child: GestureDetector(
               onTap: widget.onTooltipTap,
-              child: widget.container ?? const SizedBox.shrink(),
+              child: RepaintBoundary(
+                child: widget.container ?? const SizedBox.shrink(),
+              ),
             ),
           )
         : MouseRegion(
@@ -188,29 +192,31 @@ class _ToolTipWrapperState extends State<ToolTipWrapper>
                 : SystemMouseCursors.click,
             child: GestureDetector(
               onTap: widget.onTooltipTap,
-              child: Container(
-                padding: widget.tooltipPadding,
-                decoration: BoxDecoration(
-                  color: widget.tooltipBackgroundColor,
-                  borderRadius: widget.tooltipBorderRadius ??
-                      const BorderRadius.all(Radius.circular(8)),
-                ),
-                child: ToolTipContent(
-                  title: widget.title,
-                  description: widget.description,
-                  titleTextAlign: widget.titleTextAlign,
-                  descriptionTextAlign: widget.descriptionTextAlign,
-                  titleAlignment: widget.titleAlignment,
-                  descriptionAlignment: widget.descriptionAlignment,
-                  textColor: widget.textColor,
-                  titleTextStyle: widget.titleTextStyle,
-                  descTextStyle: widget.descTextStyle,
-                  titlePadding: widget.titlePadding,
-                  descriptionPadding: widget.descriptionPadding,
-                  titleTextDirection: widget.titleTextDirection,
-                  descriptionTextDirection: widget.descriptionTextDirection,
-                  tooltipActionConfig: widget.tooltipActionConfig,
-                  tooltipActions: widget.tooltipActions,
+              child: RepaintBoundary(
+                child: Container(
+                  padding: widget.tooltipPadding,
+                  decoration: BoxDecoration(
+                    color: widget.tooltipBackgroundColor,
+                    borderRadius: widget.tooltipBorderRadius ??
+                        const BorderRadius.all(Radius.circular(8)),
+                  ),
+                  child: ToolTipContent(
+                    title: widget.title,
+                    description: widget.description,
+                    titleTextAlign: widget.titleTextAlign,
+                    descriptionTextAlign: widget.descriptionTextAlign,
+                    titleAlignment: widget.titleAlignment,
+                    descriptionAlignment: widget.descriptionAlignment,
+                    textColor: widget.textColor,
+                    titleTextStyle: widget.titleTextStyle,
+                    descTextStyle: widget.descTextStyle,
+                    titlePadding: widget.titlePadding,
+                    descriptionPadding: widget.descriptionPadding,
+                    titleTextDirection: widget.titleTextDirection,
+                    descriptionTextDirection: widget.descriptionTextDirection,
+                    tooltipActionConfig: widget.tooltipActionConfig,
+                    tooltipActions: widget.tooltipActions,
+                  ),
                 ),
               ),
             ),
@@ -250,7 +256,12 @@ class _ToolTipWrapperState extends State<ToolTipWrapper>
           _TooltipLayoutId(
             id: TooltipLayoutSlot.tooltipBox,
             key: UniqueKey(),
-            child: defaultToolTipWidget,
+            child: !widget.semanticEnable
+                ? defaultToolTipWidget
+                : Semantics(
+                    liveRegion: true,
+                    child: defaultToolTipWidget,
+                  ),
           ),
           if (widget.tooltipActions.isNotEmpty &&
               (widget.tooltipActionConfig.position.isOutside ||